+++ src/main/java/com/powsybl/dynawo/dynamicmodels/AbstractLoadModel.java
@@ -1,0 +1,54 @@
+/**
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.dynawo.dynamicmodels;
+
+import static com.powsybl.dynawo.xml.DynawoXmlConstants.DYN_URI;
+import static com.powsybl.dynawo.xml.DynawoXmlConstants.MACRO_CONNECTOR_PREFIX;
+import static com.powsybl.dynawo.xml.DynawoXmlConstants.MACRO_STATIC_REFERENCE_PREFIX;
+import static com.powsybl.dynawo.xml.DynawoXmlConstants.NETWORK;
+
+import javax.xml.stream.XMLStreamException;
+import javax.xml.stream.XMLStreamWriter;
+
+import com.powsybl.dynawo.xml.DynawoXmlContext;
+import com.powsybl.dynawo.xml.MacroConnectorXml;
+
+/**
+ * @author Marcos de Miguel <demiguelm at aia.es>
+ */
+public abstract class AbstractLoadModel extends AbstractBlackBoxModel {
+
+    public AbstractLoadModel(String dynamicModelId, String staticId, String parameterSetId) {
+        super(dynamicModelId, staticId, parameterSetId);
+    }
+
+    @Override
+    public void write(XMLStreamWriter writer, DynawoXmlContext context) throws XMLStreamException {
+        if (context.getIndex(getLib(), true) == 0) {
+            // Write the macroStaticReference object
+            writer.writeStartElement(DYN_URI, "macroStaticReference");
+            writer.writeAttribute("id", MACRO_STATIC_REFERENCE_PREFIX + getLib());
+            writeReference(writer, context);
+            writer.writeEndElement();
+
+            // Write the macroConnector object
+            writer.writeStartElement(DYN_URI, "macroConnector");
+            writer.writeAttribute("id", MACRO_CONNECTOR_PREFIX + getLib());
+            writeConnector(writer, context);
+            writer.writeEndElement();
+        }
+
+        writeBlackBoxModel(writer, context);
+
+        // Write the connect object
+        MacroConnectorXml.writeMacroConnect(writer, MACRO_CONNECTOR_PREFIX + getLib(), getDynamicModelId(), NETWORK);
+    }
+
+    protected abstract void writeReference(XMLStreamWriter writer, DynawoXmlContext context) throws XMLStreamException;
+
+    protected abstract void writeConnector(XMLStreamWriter writer, DynawoXmlContext context) throws XMLStreamException;
+}
